Remove Favorites section header from NTP (#3381)
Task/Issue URL: https://app.asana.com/0/0/1208244619690577/f
Tech Design URL:
CC:

**Description**:

Removes Favorites header when no favorites present. This allowed to
removed whole EmptyStateView. Leftover function was incorporated into
`FavoritesViewModel`, allowing to remove the generic type from
`NewTabPageView`.

**Steps to test this PR**:
1. Open NTP without having any favorites, make sure there's an add
button and placeholders are visible.
2. Rotate to landscape. Verify whole row is filled up with placeholders.
3. Pixel should be fired when any placeholder is tapped. 

**Definition of Done (Internal Only)**:

* [x] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

**Orientation Testing**:

* [x] Portrait
* [x] Landscape
---
###### Internal references:
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
