+++ DuckDuckGo/cs.lproj/Localizable.strings
@@ -1712,9 +1712,6 @@
 /* Title for the VPN Settings screen. */
 "network.protection.vpn.settings.title" = "Nastavení sítě VPN";
 
-/* Header title of Favorites section */
-"new.tab.page.favorites.setion.header.title" = "Oblíbené";
-
 /* Information message about New Tab Page redesign */
 "new.tab.page.intro.message.body" = "Přizpůsob si oblíbené záložky a časté funkce. Přeskládej je nebo je skryj, ať je všechno přehledné.";
