+++ DuckDuckGoTests/NewTabPageFavoritesModelTests.swift
@@ -88,20 +88,48 @@ final class NewTabPageFavoritesModelTests: XCTestCase {
         XCTAssertEqual(PixelFiringMock.lastPixel, .newTabPageFavoritesPlaceholderTapped)
     }
 
-    func testFiresPixelOnShowingTooltip() {
+    func testPrefixFavoritesCreatesRemainingPlaceholders() {
         let sut = createSUT()
 
-        XCTAssertFalse(sut.isShowingTooltip)
-        sut.toggleTooltip()
+        let slice = sut.prefixedFavorites(for: 3)
 
-        XCTAssertEqual(PixelFiringMock.lastPixel, .newTabPageFavoritesInfoTooltip)
+        XCTAssertEqual(slice.items.filter(\.isPlaceholder).count, 2)
+        XCTAssertEqual(slice.items.count, 3)
+        XCTAssertFalse(slice.isCollapsible)
     }
 
-    private func createSUT() -> FavoritesDefaultViewModel {
-        FavoritesDefaultViewModel(favoriteDataSource: favoriteDataSource,
-                                  faviconLoader: FavoritesFaviconLoader(),
-                                  pixelFiring: PixelFiringMock.self,
-                                  dailyPixelFiring: PixelFiringMock.self)
+    func testPrefixFavoritesLimitsToTwoRows() {
+        favoriteDataSource.favorites.append(contentsOf: Array(repeating: Favorite.stub(), count: 10))
+        let sut = createSUT()
+
+        let slice = sut.prefixedFavorites(for: 4)
+
+        XCTAssertEqual(slice.items.count, 8)
+        XCTAssertTrue(slice.isCollapsible)
+    }
+
+    func testAddItemIsLastWhenFavoritesPresent() throws {
+        favoriteDataSource.favorites.append(contentsOf: Array(repeating: Favorite.stub(), count: 10))
+        let sut = createSUT()
+        
+        let lastItem = try XCTUnwrap(sut.allFavorites.last)
+
+        XCTAssertTrue(lastItem == .addFavorite)
+    }
+
+    func testAddItemIsFirstWhenFavoritesEmpty() throws {
+        let sut = createSUT()
+        
+        let firstItem = try XCTUnwrap(sut.allFavorites.first)
+        
+        XCTAssertTrue(firstItem == .addFavorite)
+    }
+
+    private func createSUT() -> FavoritesViewModel {
+        FavoritesViewModel(favoriteDataSource: favoriteDataSource,
+                           faviconLoader: FavoritesFaviconLoader(),
+                           pixelFiring: PixelFiringMock.self,
+                           dailyPixelFiring: PixelFiringMock.self)
     }
 }
 
@@ -127,5 +155,14 @@ private final class MockNewTabPageFavoriteDataSource: NewTabPageFavoriteDataSour
 private extension Favorite {
     static func stub() -> Favorite {
         Favorite(id: UUID().uuidString, title: "foo", domain: "bar")
+    }
+}
+
+private extension FavoriteItem {
+    var isPlaceholder: Bool {
+        switch self {
+        case .placeholder: return true
+        case .favorite, .addFavorite: return false
+        }
     }
 }
